refactor: split github function into a separate module
