+++ packages_rs/nextclade-web/src/pages/_app.tsx
@@ -41,7 +41,8 @@ import { ReactQueryDevtools } from 'react-query/devtools'
 
 import { DOMAIN_STRIPPED } from 'src/constants'
 import { parseUrl } from 'src/helpers/parseUrl'
-import { initializeDatasets, initializeGithubDataset } from 'src/io/fetchDatasets'
+import { initializeDatasets } from 'src/io/fetchDatasets'
+import { initializeGithubDataset } from 'src/io/fetchDatasetFromGithub'
 import { ErrorPopup } from 'src/components/Error/ErrorPopup'
 import Loading from 'src/components/Loading/Loading'
 import { LinkExternal } from 'src/components/Link/LinkExternal'
